requirements.txt
streamlit 
matplotlib.pyplot=3.9.0
numpy
PIL
pytesseract
